# backend/api-collections/fleet-manager/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/missions": {
      "get": {
        "operationId": "MissionController_getMissionsAsync",
        "summary": "Lists missions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter missions by name",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MissionType"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/MissionStatus"
            }
          },
          {
            "name": "droneId",
            "required": false,
            "in": "query",
            "description": "Filter missions by assigned drone ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pilotName",
            "required": false,
            "in": "query",
            "description": "Filter missions by pilot name",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter missions scheduled on or after this date (ISO 8601)",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter missions scheduled on or before this date (ISO 8601)",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of missions has been successfully retrieved."
          }
        },
        "tags": [
          "missions"
        ]
      },
      "post": {
        "operationId": "MissionController_createMissionAsync",
        "summary": "Creates a new mission",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The data for the mission to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMissionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The mission has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMissionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing data."
          }
        },
        "tags": [
          "missions"
        ]
      }
    },
    "/missions/{id}": {
      "get": {
        "operationId": "MissionController_getMissionAsync",
        "summary": "Retrieves information about a specific mission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the mission",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The mission information has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMissionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Mission not found."
          }
        },
        "tags": [
          "missions"
        ]
      },
      "put": {
        "operationId": "MissionController_updateMissionAsync",
        "summary": "Updates an existing mission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the mission to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The data for the mission to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMissionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The mission has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateMissionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or missing data."
          },
          "404": {
            "description": "Mission not found."
          }
        },
        "tags": [
          "missions"
        ]
      },
      "delete": {
        "operationId": "MissionController_softDeleteMissionAsync",
        "summary": "Deletes a specific mission (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The unique ID of the mission to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The mission has been successfully deleted."
          },
          "404": {
            "description": "Mission not found."
          }
        },
        "tags": [
          "missions"
        ]
      }
    },
    "/missions/{id}/pre-flight": {
      "patch": {
        "operationId": "MissionController_preFlightCheckMissionAsync",
        "summary": "Starts pre-flight check for a mission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "missions"
        ]
      }
    },
    "/missions/{id}/start": {
      "patch": {
        "operationId": "MissionController_startMissionAsync",
        "summary": "Starts a mission (IN_PROGRESS)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "missions"
        ]
      }
    },
    "/missions/{id}/complete": {
      "patch": {
        "operationId": "MissionController_completeMissionAsync",
        "summary": "Completes a mission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteMissionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "missions"
        ]
      }
    },
    "/missions/{id}/abort": {
      "patch": {
        "operationId": "MissionController_abortMissionAsync",
        "summary": "Aborts a mission",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbortMissionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "missions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fleet Manager API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "MissionType": {
        "type": "string",
        "description": "Type of the mission",
        "enum": [
          "WIND_TURBINE_INSPECTION",
          "SOLAR_PANEL_SURVEY",
          "POWER_LINE_PATROL"
        ]
      },
      "MissionStatus": {
        "type": "string",
        "enum": [
          "PLANNED",
          "PRE_FLIGHT_CHECK",
          "IN_PROGRESS",
          "COMPLETED",
          "ABORTED"
        ]
      },
      "GetMissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the mission"
          },
          "name": {
            "type": "string",
            "description": "Name of the mission"
          },
          "type": {
            "type": "string",
            "description": "Type of the mission",
            "enum": [
              "WIND_TURBINE_INSPECTION",
              "SOLAR_PANEL_SURVEY",
              "POWER_LINE_PATROL"
            ]
          },
          "droneId": {
            "type": "string",
            "description": "UUID of the assigned drone"
          },
          "pilotName": {
            "type": "string",
            "description": "Name of the assigned pilot"
          },
          "siteLocation": {
            "type": "string",
            "description": "Location or site of the mission"
          },
          "scheduledStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled start time of the mission"
          },
          "scheduledEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled end time of the mission"
          },
          "actualStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Actual start time of the mission"
          },
          "actualEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Actual end time of the mission"
          },
          "status": {
            "type": "string",
            "description": "Status of the mission",
            "enum": [
              "PLANNED",
              "PRE_FLIGHT_CHECK",
              "IN_PROGRESS",
              "COMPLETED",
              "ABORTED"
            ]
          },
          "flightHoursAtCompletion": {
            "type": "number",
            "description": "Flight hours logged at mission completion"
          },
          "abortReason": {
            "type": "string",
            "description": "Reason for aborting the mission, if applicable"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record update timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "droneId",
          "pilotName",
          "siteLocation",
          "scheduledStartTime",
          "scheduledEndTime",
          "actualStartTime",
          "actualEndTime",
          "status",
          "flightHoursAtCompletion",
          "abortReason",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "CreateMissionRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the mission",
            "maxLength": 255,
            "example": "Alpha Perimeter Inspection"
          },
          "type": {
            "example": "POWER_LINE_PATROL",
            "$ref": "#/components/schemas/MissionType"
          },
          "droneId": {
            "type": "string",
            "description": "UUID of the assigned drone",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "pilotName": {
            "type": "string",
            "description": "Name of the assigned pilot",
            "maxLength": 255,
            "example": "John Doe"
          },
          "siteLocation": {
            "type": "string",
            "description": "Location or site of the mission",
            "maxLength": 255,
            "example": "Site Alpha - Sector 4"
          },
          "scheduledStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled start time of the mission (ISO 8601)",
            "example": "2026-08-01T08:00:00Z"
          },
          "scheduledEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled end time of the mission (ISO 8601)",
            "example": "2026-08-01T12:00:00Z"
          }
        },
        "required": [
          "name",
          "type",
          "droneId",
          "pilotName",
          "siteLocation",
          "scheduledStartTime",
          "scheduledEndTime"
        ]
      },
      "CreateMissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the mission"
          },
          "name": {
            "type": "string",
            "description": "Name of the mission"
          },
          "type": {
            "type": "string",
            "description": "Type of the mission",
            "enum": [
              "WIND_TURBINE_INSPECTION",
              "SOLAR_PANEL_SURVEY",
              "POWER_LINE_PATROL"
            ]
          },
          "droneId": {
            "type": "string",
            "description": "UUID of the assigned drone"
          },
          "pilotName": {
            "type": "string",
            "description": "Name of the assigned pilot"
          },
          "siteLocation": {
            "type": "string",
            "description": "Location or site of the mission"
          },
          "scheduledStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled start time of the mission"
          },
          "scheduledEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled end time of the mission"
          },
          "actualStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Actual start time of the mission"
          },
          "actualEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Actual end time of the mission"
          },
          "status": {
            "type": "string",
            "description": "Status of the mission",
            "enum": [
              "PLANNED",
              "PRE_FLIGHT_CHECK",
              "IN_PROGRESS",
              "COMPLETED",
              "ABORTED"
            ]
          },
          "flightHoursAtCompletion": {
            "type": "number",
            "description": "Flight hours logged at mission completion"
          },
          "abortReason": {
            "type": "string",
            "description": "Reason for aborting the mission, if applicable"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "droneId",
          "pilotName",
          "siteLocation",
          "scheduledStartTime",
          "scheduledEndTime",
          "actualStartTime",
          "actualEndTime",
          "status",
          "flightHoursAtCompletion",
          "abortReason",
          "createdAt",
          "version"
        ]
      },
      "UpdateMissionRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the mission",
            "maxLength": 255,
            "example": "Alpha Perimeter Inspection"
          },
          "type": {
            "example": "POWER_LINE_PATROL",
            "$ref": "#/components/schemas/MissionType"
          },
          "droneId": {
            "type": "string",
            "description": "UUID of the assigned drone",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "pilotName": {
            "type": "string",
            "description": "Name of the assigned pilot",
            "maxLength": 255,
            "example": "John Doe"
          },
          "siteLocation": {
            "type": "string",
            "description": "Location or site of the mission",
            "maxLength": 255,
            "example": "Site Alpha - Sector 4"
          },
          "scheduledStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled start time of the mission (ISO 8601). Required if scheduledEndTime is provided.",
            "example": "2026-08-01T08:00:00Z"
          },
          "scheduledEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled end time of the mission (ISO 8601). Required if scheduledStartTime is provided.",
            "example": "2026-08-01T12:00:00Z"
          },
          "version": {
            "type": "number",
            "description": "Current row version for optimistic locking",
            "example": 1
          }
        },
        "required": [
          "version"
        ]
      },
      "UpdateMissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the mission"
          },
          "name": {
            "type": "string",
            "description": "Name of the mission"
          },
          "type": {
            "type": "string",
            "description": "Type of the mission",
            "enum": [
              "WIND_TURBINE_INSPECTION",
              "SOLAR_PANEL_SURVEY",
              "POWER_LINE_PATROL"
            ]
          },
          "droneId": {
            "type": "string",
            "description": "UUID of the assigned drone"
          },
          "pilotName": {
            "type": "string",
            "description": "Name of the assigned pilot"
          },
          "siteLocation": {
            "type": "string",
            "description": "Location or site of the mission"
          },
          "scheduledStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled start time of the mission"
          },
          "scheduledEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Scheduled end time of the mission"
          },
          "actualStartTime": {
            "format": "date-time",
            "type": "string",
            "description": "Actual start time of the mission"
          },
          "actualEndTime": {
            "format": "date-time",
            "type": "string",
            "description": "Actual end time of the mission"
          },
          "status": {
            "type": "string",
            "description": "Status of the mission",
            "enum": [
              "PLANNED",
              "PRE_FLIGHT_CHECK",
              "IN_PROGRESS",
              "COMPLETED",
              "ABORTED"
            ]
          },
          "flightHoursAtCompletion": {
            "type": "number",
            "description": "Flight hours logged at mission completion"
          },
          "abortReason": {
            "type": "string",
            "description": "Reason for aborting the mission, if applicable"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Record update timestamp"
          },
          "version": {
            "type": "number",
            "description": "Row version for optimistic locking"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "droneId",
          "pilotName",
          "siteLocation",
          "scheduledStartTime",
          "scheduledEndTime",
          "actualStartTime",
          "actualEndTime",
          "status",
          "flightHoursAtCompletion",
          "abortReason",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
      "CompleteMissionRequestDto": {
        "type": "object",
        "properties": {
          "flightHoursAtCompletion": {
            "type": "number",
            "description": "Flight hours logged upon completion of the mission",
            "example": 3.5
          },
          "version": {
            "type": "number",
            "description": "Current row version for optimistic locking",
            "example": 1
          }
        },
        "required": [
          "flightHoursAtCompletion",
          "version"
        ]
      },
      "AbortMissionRequestDto": {
        "type": "object",
        "properties": {
          "abortReason": {
            "type": "string",
            "description": "Reason for aborting the mission",
            "example": "Adverse weather conditions"
          },
          "flightHoursAtAborting": {
            "type": "number",
            "description": "Flight hours logged upon aborting the mission",
            "example": 3.5
          },
          "version": {
            "type": "number",
            "description": "Current row version for optimistic locking",
            "example": 1
          }
        },
        "required": [
          "abortReason",
          "flightHoursAtAborting",
          "version"
        ]
      },
...
    }
  }
}